# acju-prayer-times-docs/openapi.yaml
openapi: 3.0.3
info:
  title: ACJU Prayer Times API
  version: 1.0.0
  description: >
    REST API providing prayer times across 13 regions in Sri Lanka based on 
    prayer-time data sourced from the All Ceylon Jamiyyathul Ulama (ACJU).
    Note: This API is independently developed by KR Hasni Zihar and is not an official ACJU API.
servers:
  - url: https://salahsl.vercel.app
paths:
  /health:
    get:
      summary: Lightweight Root Health Check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
  /api/v1/metadata:
    get:
      summary: Get dataset metadata and version
      description: Returns information about the underlying ACJU dataset, verification date, and API version.
      responses:
        '200':
          description: Successful response
  /api/v1/health:
    get:
      summary: API Health Check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  database:
                    type: string
                  provider:
                    type: string
                  timezone:
                    type: string
                  available_years:
                    type: array
                    items:
                      type: integer
                  locations:
                    type: integer
                  monthly_sources:
                    type: integer
                  valid_prayer_records:
                    type: integer
                  source_anomalies:
                    type: integer
  /api/v1/locations:
    get:
      summary: List all active locations
      responses:
        '200':
          description: Successful response
  /api/v1/locations/{slug}:
    get:
      summary: Get location by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Location not found
  /api/v1/locations/resolve:
    get:
      summary: Resolve GPS coordinates to an ACJU region
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
      responses:
        '200':
          description: Successfully resolved to an ACJU location
        '400':
          description: Missing or invalid coordinates, or outside Sri Lanka
        '404':
          description: Location not mapped
  /api/v1/prayer-times/today/all:
    get:
      summary: Get today's valid prayer times for all 13 locations
      description: Returns times for today based on the Asia/Colombo timezone. Excludes anomalous source dates.
      responses:
        '200':
          description: Successful response
        '404':
          description: Data not available
  /api/v1/prayer-times/today:
    get:
      summary: Get today's prayer times for a specific location
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
          description: The slug of the location (e.g., 'batticaloa-ampara'). Required if lat/lng are not provided.
        - name: lat
          in: query
          required: false
          schema:
            type: number
        - name: lng
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful response
        '400':
          description: Missing location/coordinates or outside Sri Lanka
        '404':
          description: Location or data not found
  /api/v1/prayer-times/{location}/{date}:
    get:
      summary: Get prayer times for a specific valid calendar date
      parameters:
        - name: location
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Format YYYY-MM-DD
      responses:
        '200':
          description: Successful response
        '400':
          description: Invalid date format
        '404':
          description: Data not available (e.g. for anomaly dates like 2026-02-29, this returns 404 since it's not a valid calendar date)
  /api/v1/prayer-times/{location}/{year}/{month}:
    get:
      summary: Get all prayer times for a specific month
      parameters:
        - name: location
          in: path
          required: true
          schema:
            type: string
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
        '400':
          description: Invalid year or month
        '404':
          description: Data not available
  /api/v1/prayer-times/{location}:
    get:
      summary: Get prayer times within a date range
      parameters:
        - name: location
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
        '400':
          description: Invalid range or date format
        '404':
          description: Data not available